feat(files): FileTreePanel 解耦 + 像素精修 (v1-08)
FileTreePanel.tsx 从 1,399 行拆分至 9 个子文件，每个 ≤300 行：
- fileTreeTypes.ts (41 行): 类型定义
- fileTreeHelpers.ts (270 行): 纯函数 (buildTreeSnapshot, flattenTree, etc.)
- useFileTreeKeyboard.ts (139 行): 键盘导航 handleTreeKeyDown
- fileTreeContextMenu.ts (86 行): 右键菜单 buildNodeContextMenuItems
- FileTreeRenameRow.tsx (88 行): 重命名行组件
- FileTreeNodeRow.tsx (300 行): 节点行组件 (含拖拽、上下文菜单、展开/折叠)
- useFileTreeCore.ts (145 行): 核心状态 + 副作用
- useFileTreeState.ts (151 行): CRUD handlers + 组合层
- FileTreePanel.tsx (126 行): 薄壳 (布局 + 子组件编排)

视觉精修 (AC-1~7):
- AC-1: IconBar 固定 w-12 shrink-0 (48px)
- AC-2: 树节点行高 h-8 (32px)
- AC-3: 拖拽指示器增加 8px 圆形手柄
- AC-4: Rename input focus 边框 --color-info
- AC-5: ContextMenu 支持 shortcut 字段渲染 (11px, --color-fg-muted)
- AC-7: 展开/折叠箭头 CSS rotation + transition-transform

42/42 测试通过 (19 原有 + 23 守卫)
全量回归 2353 tests pass | typecheck clean | lint 0 errors

Closes #1191
